+++ src/Contracts/Factories/FactoryInterface.php
@@ -21,8 +21,7 @@ namespace CloudCreativity\JsonApi\Contracts\Factories;
 use CloudCreativity\JsonApi\Contracts\ContainerInterface;
 use CloudCreativity\JsonApi\Contracts\Encoder\SerializerInterface;
 use CloudCreativity\JsonApi\Contracts\Http\Client\ClientInterface;
-use CloudCreativity\JsonApi\Contracts\Http\Requests\RequestInterface;
-use CloudCreativity\JsonApi\Contracts\Http\Requests\RequestInterpreterInterface;
+use CloudCreativity\JsonApi\Contracts\Http\Requests\InboundRequestInterface;
 use CloudCreativity\JsonApi\Contracts\Http\Responses\ResponseInterface;
 use CloudCreativity\JsonApi\Contracts\Object\DocumentInterface;
 use CloudCreativity\JsonApi\Contracts\Pagination\PageInterface;
@@ -34,6 +33,7 @@ use CloudCreativity\JsonApi\Contracts\Validators\ValidatorFactoryInterface;
 use Neomerx\JsonApi\Contracts\Codec\CodecMatcherInterface;
 use Neomerx\JsonApi\Contracts\Document\ErrorInterface;
 use Neomerx\JsonApi\Contracts\Document\LinkInterface;
+use Neomerx\JsonApi\Contracts\Encoder\Parameters\EncodingParametersInterface;
 use Neomerx\JsonApi\Contracts\Factories\FactoryInterface as BaseFactoryInterface;
 use Neomerx\JsonApi\Contracts\Http\Query\QueryCheckerInterface;
 use Neomerx\JsonApi\Contracts\Schema\ContainerInterface as SchemaContainerInterface;
@@ -41,7 +41,6 @@ use Neomerx\JsonApi\Encoder\EncoderOptions;
 use Neomerx\JsonApi\Exceptions\ErrorCollection;
 use Psr\Http\Message\RequestInterface as PsrRequest;
 use Psr\Http\Message\ResponseInterface as PsrResponse;
-use Psr\Http\Message\ServerRequestInterface;
 
 /**
  * Interface FactoryInterface
@@ -62,20 +61,25 @@ interface FactoryInterface extends BaseFactoryInterface
     public function createSerializer(SchemaContainerInterface $container, EncoderOptions $encoderOptions = null);
 
     /**
-     * Create a JSON API request object from a PSR server request.
+     * Create an inbound request object from the supplied parameters.
      *
-     * @param ServerRequestInterface $httpRequest
-     *      the inbound HTTP request
-     * @param RequestInterpreterInterface $interpreter
-     *      the interpreter to analyze the request
-     * @param StoreInterface $store
-     *      the store that the request relates to.
-     * @return RequestInterface
+     * @param $method
+     * @param $resourceType
+     * @param null $resourceId
+     * @param null $relationshipName
+     * @param bool $relationships
+     * @param DocumentInterface|null $document
+     * @param EncodingParametersInterface|null $parameters
+     * @return InboundRequestInterface
      */
-    public function createRequest(
-        ServerRequestInterface $httpRequest,
-        RequestInterpreterInterface $interpreter,
-        StoreInterface $store
+    public function createInboundRequest(
+        $method,
+        $resourceType,
+        $resourceId = null,
+        $relationshipName = null,
+        $relationships = false,
+        DocumentInterface $document = null,
+        EncodingParametersInterface $parameters = null
     );
 
     /**